Refactor: TripsService to use Mongoose trip model
Updated the TripsService to replace the custom repository with the Mongoose trip model implementation. This change enhances integration with MongoDB and leverages Mongoose's built-in functionalities. Corresponding unit tests were updated to reflect these changes.
